feat: handle unread notifications
